pyproject.toml
[project]
name = "accounts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cyclopts>=3.11.1",
    "duckdb>=1.2.1",
    "fredapi>=0.5.2",
    "kaleido==0.2.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "nicegui>=2.12.1",
    "numpy-financial>=1.0.0",
    "orjson>=3.10.16",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "prefixed>=0.9.0",
    "psutil>=7.0.0",
    "pyecharts>=2.0.8",
    "pyngleton>=0.0.6",
    "pytest-playwright>=0.7.0",
    "redis>=5.2.1",
    "statsmodels>=0.14.4",
    "textual>=2.1.2",
    "textual-dev>=1.7.0",
    "walrus>=0.9.4",
    "yahoofinancials>=1.20",
    "yahooquery>=2.3.7",
    "yfinance>=0.2.54",
]

[tool.uv]
dev-dependencies = [
    "pandas-stubs>=2.2.2.240909",
    "toml>=0.10.2",
]
